requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
tensorflow==2.16.2